+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Change log
 - Changed the code to raise a ``DisallowedHost`` exception instead of a 404 if
   unable to find a matching host for the current request.
 - Switched to pre-commit.
+- Fixed a bug where the ``redirect_to_site_middleware`` would redirect too
+  often.
 
 
 `0.0.1`_ (2021-09-14)